move getDateErrors
