+++ app/pages/sections/catalogue-solutions/order-item/contextCreator.js
@@ -1,6 +1,6 @@
 import { baseUrl } from '../../../../config';
 import { generateErrorMap } from '../../../../helpers/contextCreators/generateErrorMap';
-import { generateQuestions } from '../../../../helpers/generateQuestions';
+import { generateQuestions } from '../../../../helpers/contextCreators/generateQuestions';
 import { generateErrorSummary } from '../../../../helpers/contextCreators/generateErrorSummary';
 import { generateAddPriceTable } from '../../../../helpers/contextCreators/generateAddPriceTable';
