+++ app/pages/sections/commencement-date/controller.js
@@ -1,6 +1,6 @@
 import { getData, putData } from 'buying-catalogue-library';
 import { getContext, getErrorContext } from './contextCreator';
-import { getDateErrors } from '../../../helpers/getDateErrors';
+import { getDateErrors } from '../../../helpers/controllers/getDateErrors';
 import { getEndpoint } from '../../../endpoints';
 import { logger } from '../../../logger';
 import { extractDate } from '../../../helpers/controllers/extractDate';
